requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-admin-thumbnails==0.2.6
django-ckeditor==6.2.0
django-filter==21.1
django-jalali==6.0.0
django-js-asset==2.0.0
django-phone-field==1.8.1
django-taggit==2.1.0
idna==3.3
jdatetime==4.1.0
Pillow==9.0.1
psycopg2==2.9.3
requests==2.27.1
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
suds-py3==1.4.5.0
tzdata==2021.5
urllib3==1.26.8
